[Core] Moved test drivers from CoreBundle/Tests/ to CoreBundle/Test/
