+++ app/config/config_test.yml
@@ -20,7 +20,7 @@ doctrine:
     dbal:
         dbname: %sylius.database.name%_test
         path:   %sylius.database.path%
-        driver_class: Sylius\Bundle\CoreBundle\Tests\MySqlDriver
+        driver_class: Sylius\Bundle\CoreBundle\Test\MySqlDriver
 
 sylius_money:
     currency: EUR
